Move Voice complexity to Chorus
